Disabled all LazyExtras lang imports
There was a problem with a large code-base like Flink. Nvim --embed
consumed 100% CPU on one core and the UI was unresponsive.

I did step-by-step disable all plugins and figured out that it seems
like the cause was that the LSP:s configured by LazyExtras somehow
conflicted (or had a different config) than the Java and Scala LSPs
explicitly configured.

* disabled all "lang" plugins in lazy.lua.
* added a "disabled.lua" for easy disabling of plugins.
* lsp-metals conditionally load "cmp_nvim_lsp" so that it won't crash if
  that plug is disabled.